refactor: divide classes between packages, add factories
