+++ yggdrasil-cartago/src/main/java/org/hyperagents/yggdrasil/cartago/entities/HypermediaInterfaceConstructor.java
@@ -1,4 +1,4 @@
-package org.hyperagents.yggdrasil.cartago;
+package org.hyperagents.yggdrasil.cartago.entities;
 
 import cartago.ArtifactDescriptor;
 import cartago.ArtifactId;
